[server] Metric Name fix and math fix for HeartbeatMonitor (#869)
* [server] Metric Name fix and math fix for HeartbeatMonitor

The resulting metric name contained an extra set of '--' when being hooked up to the reporter.  Additionally, the metric was reporting the time of the last heartbeat, but not the time SINCE the last heartbeat.

+++ clients/da-vinci-client/src/main/java/com/linkedin/davinci/stats/ingestion/heartbeat/HeartbeatMonitoringService.java
@@ -230,7 +230,7 @@ public class HeartbeatMonitoringService extends AbstractVeniceService {
 
   @FunctionalInterface
   interface ReportLagFunction {
-    void apply(String storeName, int version, String region, double lag);
+    void apply(String storeName, int version, String region, long lag);
   }
 
   private class HeartbeatReporterThread extends Thread {
